pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "work-agents"
version = "0.1.0"
description = "Multi-agent software team orchestration system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jira>=3.8.0",
    "PyGithub>=2.3.0",
    "gitpython>=3.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "rich>=13.9.0",
    "click>=8.1.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "respx>=0.21.0",
]

[project.scripts]
work-agents = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
